BROKEN: mixer_paths: Import mmap-playback paths from stock
This is broken because these paths configure the secondary (MM_NOIRQ_2)
DAI, as the first is configured (on stock) through ULL. That is wrong
(doesn't work on SODP, likely because the HAL requires modifications) so
instead a followup commit will "paste" the stock audio-ull-playback ctls
over these mmap-playback paths.

Note that a matching change has been made in audio_platform_info to make
sure the right PCM (17 for MM_NOIRQ, instead of 33 for MM_NOIRQ_2) is
selected.

Signed-off-by: Marijn Suijten <marijns95@gmail.com>