mixer_paths: Convert mmap-playback to use the primary MM_NOIRQ DAI
As explained in the previous commit stock uses the primary DAI for ULL
and the secondary for mmap. This is wrong on SODP, instead we use the
ULL DAI for... audio-ull-playback... and the MM_NOIRQ DAI for
mmap-playback. Note that this requires a similar change in
audio_platform_info to select the right PCM for the _ULL and _MMAP
usecase respectively.

Signed-off-by: Marijn Suijten <marijns95@gmail.com>